requirements.txt
mangum==0.19.0
fastapi==0.115.6
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.20
requests==2.32.3
geopy==2.4.1
Pillow==11.0.0
numpy==2.2.1
httpx==0.28.1
email-validator==2.2.0
aiofiles==24.1.0
psutil==5.9.8
sqlalchemy==2.0.36
